Add Non-Parametric Bootstrap Estimation of Parameter Estimate Standard Errors
This updated adds non-parametric bootstrap functionality to estimate the
SE of  a model's parameter estimates. For each iteration defined by the
user, the ROC Toolbox will randomly sample with replacement from the
trials of a subject (holding constant condition and trial type), and
then refit the model to this new sample. Any fits to bootstrapped data
that produces a warning (e.g., singular matrix) or a bad exit flag are
discarded, and a different randomization is selected.

This update also removes calculation of the SEs of the parameter
estimates derived from the Hessian matrix due to known issues with the
Hessians obtained from the FMINCON routine.