Add Alpha Scripts to Compute Posterior
A new function calc_likelihood.m was added to in alpha, which computes
the likelihood of the model given a set of parameters. This is in alpha
because it is not necessary for MLE since the toolbox uses the
log-likelihood. The calc_likelihood function is needed for computing the
posterior.

Additionally, a script was added to demonstrate how to estimate the
posterior using some functions in the toolbox and in Matlab. More work
is needed to fully integrate this into the ROC Toolbox such that it
provides an easy interface for users of this toolbox.