Update tutorial1 and roc_solver files, and add roc_version
Updated tutorial1_script.m to include an example of using the 'bootIter'
option of the roc_solver.

Updated roc_solver to fix a very, very, very minor bug when displaying
the completed iterations.

Added a function roc_version to display the current version of the
roc_toolbox, which will be useful for those who reference the toolbox.